+++ R/pdb.R
@@ -22,13 +22,14 @@
 #' \code{is.pdb} returns TRUE if x is an object of class \sQuote{pdb} and FALSE otherwise.
 #' 
 #' @param atoms a data.frame of class \code{atoms} containing ATOM and HETATM records use to create the \code{pdb} object.
-#' @param cryst1 a list of class \code{cryst1} containing the periodical bondary conditions and space group use to create the \code{pdb} object.
+#' @param crystal a list of class \code{cryst1} containing the periodical boundary conditions and space group used to create the \code{pdb} object.
 #' @param conect a data.frame of class \code{conect} containing CONECT records use to create the \code{pdb} object.
 #' @param remark a character vector containing some REMARK records to be added to the \code{pdb} object.
 #' @param title a character vector containing some TITLE records to be added to the \code{pdb} object.
 #' @param resolution numeric value specifying the resolution; the unit should be specified as an attribute.
 #' @param x an R object to be tested.
 #' @param \dots further arguments passed to or from other methods.
+#' @param cryst1 will be deprecated and replaced by argument crystal.
 #' 
 #' @seealso 
 #' \code{\link{atoms}}, \code{\link{coords}}, \code{\link{cryst1}}, \code{\link{conect}} and \code{\link{read.pdb}}
@@ -54,14 +55,25 @@ pdb <- function(...)
 
 #' @rdname pdb
 #' @export
-pdb.default <- function(atoms, cryst1 = NULL, conect = NULL, remark = NULL, title = NULL,
-		resolution = NULL, ...)
+pdb.default <- function(atoms, crystal = NULL, conect = NULL, remark = NULL, title = NULL,
+		resolution = NULL, ..., cryst1 = NULL)
 {
-  if(missing(atoms)) stop("Please specify at least an 'atoms' object")
-  if(!is.atoms(atoms)) stop("'atoms' must be an object of class 'atoms'")
-  
-  if( ! is.null(cryst1) & ! is.cryst1(cryst1)) stop("'cryst1' must be an object of class 'cryst1'")
-  if( ! is.null(conect) & ! is.conect(conect)) stop("'conect' must be an object of class 'conect'")
+	if(missing(atoms)) stop("Please specify at least an 'atoms' object")
+	if( ! is.atoms(atoms)) stop("'atoms' must be an object of class 'atoms'")
+	
+	# Crystal cell:
+	if( ! is.null(cryst1)) {
+		if(is.null(crystal)) {
+			crystal = cryst1;
+		} else {
+			stop("Please specify only crystal!",
+				"Note: arg cryst1 will be deprecated!");
+		}
+	}
+	if( ! is.null(crystal) & ! is.crystal(crystal))
+		stop("'crystal' must be an object of class 'cryst1'");
+	if( ! is.null(conect) & ! is.conect(conect))
+		stop("'conect' must be an object of class 'conect'");
   
   if(is.list(title ) | ! is.null(dim(title ))) stop("'title' must be a vector of character strings")
   if(is.list(remark) | ! is.null(dim(remark))) stop("'remark' must be a vector of character strings")
@@ -69,7 +81,7 @@ pdb.default <- function(atoms, cryst1 = NULL, conect = NULL, remark = NULL, titl
   if(! is.character(title ) & ! is.null(title )) title  <- as.character(title )
   if(! is.character(remark) & ! is.null(remark)) remark <- as.character(remark)
   
-  to.return <- list(title = title, remark = remark, cryst1 = cryst1, atoms = atoms, conect = conect)
+  to.return <- list(title = title, remark = remark, cryst1 = crystal, atoms = atoms, conect = conect)
   if( ! is.null(resolution)) to.return$Resolution = resolution;
   
   class(to.return) <- c("pdb", "list")